requirements.txt
numpy
pytesseract
requests
cv2
Pillow
flask
